Allow user to define custom states (#155)
* fix tutor instruction about meow-append

meow-append positiones the cursor to the end of the selected region.

* fix meow--enable to avoid executing symbol-name w.r.t lambda func. (#153)

In `meow--enable`, it checks whether "self-insert" is assigned to key "a" or not
by using `symbol-name`.
However, an error occurs when key "a" is assigned to a lambda function, such as
neotree.
To avoid this problem, the assigned function is checked for symbol before `symbol-name` is executed.

* avoiding executing symbol-name on lambda in keypad

* Fix beacon symbol returned in meow--current-state (#154)

* Generalized macro to produce define-key helpers

Defines a macro meow-generate-define-key which produces a helper function
meow-STATE-define-key. This macro is then applied to all keymaps.
This commit also introduces the variable meow-keymap-alist to keep track of
states to keymaps.

* Generalize state disabling

Ensure that when normal, insert, and beacon modes are entered, they disable all
modes that may conflict with them, including custom modes. We also add
meow-state-alist to keep track of all modes.

* Generalize meow--update-cursor

Use a list of conditions instead of a long cond statement.
This ensures you can easily add your own conditions and functions.

* Generalize meow--switch-state

Modify meow--switch-state such that if the state being switched to is not
caught by the case statement, it is looked up in an alist of custom modes.

* Generalize meow--current-state

This uses the previously defined variable meow-state-alist to generalize the
meow--current-state function.

* Add function to define custom state

Please see docstrings. Adds several macros to help the user easily define
custom states.

* Autoload meow-define-state to prevent errors on startup

* Fix entry function macro and add to meow-define-state

* Make keymap generator take customization options

Suppression option and sparsity options added

* Improve variable declarations

Fixed docstrings and made indentation look nicer. Also,
removes meow-custom-state-alist in favor of two alists that should always
mirror each other. Adds meow--register-state to update both lists at once.

* Change meow-escape-or-normal-modal to always work

Previously, it only switched you to normal if you were
either in insert mode or in fundamental mode.

* Indenting consistency

My indenter got mad

* Fix wrong variable usage

* Generalize meow--render-indicator

Custom states are now added to meow-replace-state-name-list.
meow--render-indicator is rewritten to be much shorter.

* Documentation

* Rewrite meow--switch-state and extract updating to init functions.

* Change macros to functions and remove evals

* Remove entry function

* Make meow-keymap-alist use symbol keys

* Rewrite cursor updating

Remove code blocks from meow-update-cursor-functions-alist and
make defuns for each one.

* Make meow--define-state-keymap accept an override keymap

* Documentation

* Disable normal mode before saving origin commands in motion

* Variable docstring tweak

Changes variable documentation for meow-keymap-alist and
meow-update-cursor-functions-alist.

* Make keypad mode remember custom states.

* Rewrite meow--motion-init to save origin command properly

Co-authored-by: Domen Skamlic <domen@skamlic.com>
Co-authored-by: Naoki Sakamoto <29372455+naoking158@users.noreply.github.com>
Co-authored-by: tianshu <doglooksgood@gmail.com>