Fix kbd macro wrapping bug
Fixes #163. Keyboard macro was modified on beacon-insert exit to
include entering insert and escaping. However, if the keyboard macro
was a string that included a special meta-code, it would get converted
to an incorrect character when converted to a vector. This is a
workaround to insert and escape without touching the macro.