keep track of direction for no-region cases
Introduces the variable meow--backward-p intended to be used as a
backup source of direction information when there is no region. If
there is a region, previous behavior is preserved (we check the
position of the point and mark).

In this commit, meow-reverse no longer supports
meow-selection-command-fallback. This is because it is now possible to
reverse without a region.

Co-authored-by: Fredrik Bergroth <fbergroth@gmail.com>