Clarify docstring for meow-visit #431
Default behavior now is you either select one candidate (word or
symbol in buffer), or search for raw string. The regexp quote will
prevent it to be treated as regexp. Docstring should not give an
impression that `meow-visit` searches for regexps.

The dual function of `meow-visit-sanitize-completion` is also
explained in better detail.