Add advice `meow--prepare-face` to `enable-theme` instead of `load-theme` (#488)
* Add advice `meow--prepare-face` to `enable-theme` instead of `load-theme`

Add advice `meow--prepare-face` to function `enable-theme` instead of
`load-theme`. This will help prepare faces if a theme is enabled with
`enable-theme`. The modification does not change the current behavior,
because `enable-theme` is called at the end of `load-theme`, and calling
`load-theme` also triggers advice on `enable-theme`.

* Use a wrapper of `meow--prepare-face` as an advice to `enable-theme`

Adding advice `meow--prepare-face` to `enable-theme` instead of `load-theme`
(commit 64bdbee) causes `meow--prepare-face` being called twice when a theme
is enabled. This commit uses a wrapper of `meow--prepare-face` as the
advice. The wrapper calls `meow--prepare-face` only when the theme to
enable is `user`.