fix meow-esc in emacsclient -t
fixes #514. If a emacs server is started from graphical, meow-esc-mode
is never called. Therefore, meow-init-esc does not get added to the
make frame hook, and it never gets enabled for emacsclient -t.

This commit always enables meow-esc-mode, and checks whether we are
graphical inside the init and deinit functions instead of in the
global init/deinit functions.