Consider local keybindings when moving commands for the Motion state. (#517)
Instead of directly binding `H-j` to what was bound to `j`, for example, bind
`H-j` to an interactive lambda which first checks whether there is a local
binding for `j`.  If there is, use that.  If there is not, then call the orignal
command bound to `j`.

This should fix using the overrides in Magit buffers (see #493), where the
command bound to `H-k` seems to be based on where point is located when
`meow--save-origin-commands` is run.

Co-authored-by: okamsn <okamsn@users.noreply.github.com>