Add a shim for Macrostep.
Macrostep shows macro expansions using overlays either in a separate buffer or at
the location where the macro is used.  In the overlay, it binds a local map to
further expand or collapse macros.

This change takes inspiration for the Beacon state and checks whether point is
inside one of the overlays in `post-command-hook`.  However, at least one of the
commands to collapse the overlay does not seem to trigger `post-command-hook`,
so switching back to the previous state is also done during
`macrostep-mode-hook`.